I hope it's final

+++ src/main/java/com/tessnd/games_assets/file/FileService.java
@@ -56,6 +56,10 @@ public class FileService {
 
     public void delete(String fileName) throws IOException {
         Path filePath = Paths.get(fileName);
-        Files.delete(filePath);
+        if (Files.exists(filePath)) {
+            Files.delete(filePath);
+        } else {
+            throw new RuntimeException("File not found: " + fileName);
+        }
     }
 }
